d3dvideosink: fix scalling with some drivers
The back buffer should have the same size as the rendering window
so scalling is done in StretchRect from the offscreen surface to
the back buffer using the bilinear filter. If not, scalling
happens from the backbuffer to the rendering window outside the
sink, which is not correctly filtered with some drivers.