hlsdemux: don't cancel current fragment download when going to PAUSED
When going to PAUSED we prefer finishing the current fragment download
and then gracefully stop the update task. Indeed if the fragment download
is cancelled, when resuming to PLAYING we will start downloading the next
fragment potentially introducing a big gap in the MPEG TS stream which
could confuse the demuxer downstream and produce a frozen output for a
while even for short pause/play sequences.