Use the correct format when doing direct rendering
On QCOM based decoders the returned color format for a decoder
created with a surface for direct rendering is a valid number.
It uses YUV420 so we can not pass that format directly to the
base class or it will use the software based caps.
So we force the format to be the private so the base class sets the
correct caps (video/x-amc).