Calculate timestamps when converting from bytestream to avc
When the parser is used to convert a bytestream formatted media into
AVC and content is interlaced the fields have to be pushed as
independent samples according to ISO/IEC 14496‐10 spec.

With this change is ensured that each generated buffer have a PTS
which is calculated for the second field.